Code review fixes

+++ src/common.h
@@ -54,6 +54,6 @@ status_t send_wrapper (int socket_fd, struct iovec *vec, int iovcnt);
 status_t recv_wrapper_nonblock (int socket_fd, void *buf, int len, int flags,
                                 int *total);
 status_t send_wrapper_nonblock (int socket_fd, struct iovec *vec, int iovcnt,
-                                int *total);
+                                size_t *total);
 
 #endif // COMMON_H
